perf(Dokan): Divide big file
